unified opt layer implementation
