+++ cognify/optimizer/core/.gitignore
@@ -1,0 +1,2 @@
+unified_layer_opt.py
+upper_layer.py
